// pslpwcl/src/app/modules/pslpapi/model/esperienzaProfessionale.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Decodifica } from './decodifica';
import { LavEsperienzaLavoro } from './lavEsperienzaLavoro';

export interface EsperienzaProfessionale { 
    dataAssunzione?: Date;
    dataCessazione?: Date;
    flgForma?: Decodifica;
    flgLuogoLavoroItalia?: string;
    flgPT?: string;
    idEsperienzaProfessionale?: number;
    lavEsperienzaLavoro?: LavEsperienzaLavoro;
    tipoEsperienza?: Decodifica;
}