// pslpwcl/src/app/modules/pslpapi/model/modalitaLavoro.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ModalitaLavoro { 
    codModalitaLavoroMin?: string;
    dataFine?: Date;
    dataInizio?: Date;
    descrModalitaLavoro?: string;
    idSilTModalitaLavoro?: number;
    idSilTModalitaLavoroNew?: number;
}